Change out most of the entire tech stack so that websockets work
The client was changed to use `gloo` (even though that change wasn't required) for future convenience.

The server changes from a basic HTTP server where we define all the implementation (with which it was difficult to integrate a websocket endpoint that could maintain a connection while messages were sent) to a proper web framework that supports websockets more easily.