Have all pages check whether they have a user or not
Because you can start on any page with the router, we want to ensure that, if a user ends up on the page, they have the login context that allows that page.
Because it doesn't need to be a separate page from the default now, we also move the event selection page (the starter page for all registered users) to be the root page instead of a page to which the user gets redirected.