Refactor the big select to improve both code organization and editing
One of the issues with stuffing a bunch of code in the select! block is that a lot of tools don't work as well inside macros. Because of this, it's good to take code out of the macros where possible and split it into separate functions which are outside the macro and can receive the full support of tooling available.